[ADD] authorized delete route

+++ app/Http/Controllers/CommentController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Comment;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
@@ -14,13 +15,18 @@ class CommentController extends Controller
         $comment = new Comment();
         $comment->content = request()->input('content');
         $comment->article_id = request()->article_id;
+        $comment->user_id = auth()->user()->id;
         $comment->save();
         return back();
     }
     
     public function delete($id){
         $comment = Comment::find($id);
-        $comment->delete();
-        return back();
+        if(Gate::allows('comment-delete',$comment)){
+            $comment->delete();
+            return back();
+        }else{
+            return back()->with('error','Unauthorize');
+        }
     }
 }
